Update bot configuration and add ReadyListener
The commit modifies several aspects of the bot's configuration and behavior. It changes the embed color and removes the copyright disclaimer in the Config file. A new ReadyListener has been added to log details when the bot is ready and connected. Changes in database queries and main class have also been applied. Various aspects of bot's cache settings in Bot.kt were updated to improve performance.